Adapted the OLED library
The OLED Library now has a OLED class that has a display instance of the adafruit library.

This class is of the singleton pattern and abstracts the configuration away from the caller.

That way simple adafruit functions can be called with including the library but it does not need to be initialised because that is within the singleton structure.